--- java/dataImporter/.gitignore
@@ -1,2 +1,0 @@
-logs/
-target/
