--- java/dataImporter/src/main/resources/log4j.xml
@@ -1,26 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE log4j:configuration SYSTEM "log4j.dtd">
-<log4j:configuration>
-    <appender name="myConsole" class="org.apache.log4j.ConsoleAppender">
-        <layout class="org.apache.log4j.PatternLayout">
-            <param name="ConversionPattern" value="[%d{dd HH:mm:ss,SSS\} %-5p] [%t] %c{2\}:%L - %m%n"/>
-        </layout>
-        <!--过滤器设置输出的级别-->
-        <filter class="org.apache.log4j.varia.LevelRangeFilter">
-            <param name="levelMin" value="debug"/>
-            <param name="levelMax" value="error"/>
-            <param name="AcceptOnMatch" value="true"/>
-        </filter>
-    </appender>
-    <!-- 指定logger的设置，additivity指示是否遵循缺省的继承机制-->
-    <logger name="com._4paradigm.dataimporter" additivity="false">
-        <level value="warn"/>
-        <appender-ref ref="myConsole"/>
-    </logger>
-
-    <!-- 根logger的设置-->
-    <root>
-        <priority value="warn"/>
-        <appender-ref ref="myConsole"/>
-    </root>
-</log4j:configuration>
