+++ java/openmldb-import/src/test/java/com/_4paradigm/openmldb/importer/WildcardURITest.java
@@ -1,0 +1,51 @@
+/*
+ * Copyright 2021 4Paradigm
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com._4paradigm.openmldb.importer;
+
+import org.junit.Assert;
+import org.junit.Test;
+
+public class WildcardURITest {
+
+    @Test
+    public void test() {
+        String path = "hdfs://host/testdata/20180[8-9]*";
+        WildcardURI wildcardURI = new WildcardURI(path);
+        Assert.assertEquals("/testdata/20180[8-9]*", wildcardURI.getPath());
+
+        path = "hdfs://host/testdata/2018+ 0[8-9]*";
+        wildcardURI = new WildcardURI(path);
+        Assert.assertEquals("/testdata/2018+ 0[8-9]*", wildcardURI.getPath());
+
+        path = "hdfs://host/testdata/2018-01-01 00%3A00%3A00";
+        wildcardURI = new WildcardURI(path);
+        Assert.assertEquals("/testdata/2018-01-01 00%3A00%3A00", wildcardURI.getPath());
+
+        path = "hdfs://host/testdata/2018-01-01   00*";
+        wildcardURI = new WildcardURI(path);
+        Assert.assertEquals("/testdata/2018-01-01   00*", wildcardURI.getPath());
+
+        path = "hdfs://host/testdata/2018-01-01#123#";
+        wildcardURI = new WildcardURI(path);
+        Assert.assertEquals("/testdata/2018-01-01#123#", wildcardURI.getPath());
+
+        path = "hdfs://host/testdata/2018-01-01#123 +#*";
+        wildcardURI = new WildcardURI(path);
+        Assert.assertEquals("/testdata/2018-01-01#123 +#*", wildcardURI.getPath());
+    }
+
+}
